requirements.txt
pandas
rtpt
scikit-learn
matplotlib
tqdm
opencv-python
git+https://github.com/openai/CLIP.git
clip-retrieval
open_clip_torch
ipywidgets
seaborn
autokeras
pytorch-lightning
lightning-bolts
torchmetrics
wandb
sklearn